move out cascade template class
